+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * #21 Tests: DataChannel to quote SQL identifiers by default
 * #22 Tests: DatabaseChannel should log prepared statement bindings
 * #23 jdbc-test: a simple data manager to keep table definitions and cleans up the DB before each test
+* #24 bootique-jdbc-test: Table.insert explicit binding style
 
 ## 0.12